+++ tket/src/Predicates/PassGenerators.cpp
@@ -354,7 +354,7 @@ PassPtr gen_clifford_push_through_pass() {
   return std::make_shared<StandardPass>(precons, t, pc, j);
 }
 
-PassPtr gen_flatten_relabel_registers_pass(const std::string& label) {
+PassPtr gen_flatten_relabel_registers_pass(const std::string& label, bool relabel_classical_expressions) {
   Transform t =
       Transform([=](Circuit& circuit, std::shared_ptr<unit_bimaps_t> maps) {
         unsigned n_qubits = circuit.n_qubits();
@@ -366,7 +366,7 @@ PassPtr gen_flatten_relabel_registers_pass(const std::string& label) {
           relabelling_map.insert({all_qubits[i], Qubit(label, i)});
         }
 
-        circuit.rename_units(relabelling_map);
+        circuit.rename_units(relabelling_map, relabel_classical_expressions);
         changed |= update_maps(maps, relabelling_map, relabelling_map);
         return changed;
       });
@@ -376,6 +376,7 @@ PassPtr gen_flatten_relabel_registers_pass(const std::string& label) {
   nlohmann::json j;
   j["name"] = "FlattenRelabelRegistersPass";
   j["label"] = label;
+  j["relabel_classical_expressions"] = relabel_classical_expressions;
   return std::make_shared<StandardPass>(precons, t, postcons, j);
 }